backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# PDF Processing
PyPDF2==3.0.1

# Machine Learning & NLP
# Note: On Windows, if scikit-learn installation fails, use: pip install scikit-learn --only-binary :all:
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
scikit-learn>=1.3.0,<1.4.0

# Database
supabase==2.3.0

# CORS is handled by FastAPI's built-in CORSMiddleware

